requirements/base.txt
Django==1.11.5
Pillow==4.2.1
tqdm==4.15.0
bpython==0.16
django-model-utils==3.0.0


